refactor: extract shared UI rendering patterns
- Extract `toggleInSet` to dom.js as canonical location; re-export from
  flow-view-helpers.js for backward compatibility. Use it in flow-view.js,
  tab-color-filter.js, and flow-modal.js instead of inline set toggle patterns.

- Extract `toggleCollapsible` to dom.js for the repeated classList.toggle
  + chevron text update pattern. Use it in file-tree-section-dom.js.

- Rename `createSettingsItem` to `createListItem` in settings-section-builder.js
  (generic name for the list-item factory). Keep `createSettingsItem` as a
  deprecated alias for backward compatibility. Migrate settings-appearance,
  settings-configs, and settings-keybindings to use `createListItem`.

Closes #530
