+++ src/components/flow-view.js
@@ -1,6 +1,6 @@
 import { registerComponent, getComponent } from '../utils/component-registry.js';
 import { ComponentBase } from '../utils/component-base.js';
-import { moveFlowInOrder } from '../utils/flow-view-helpers.js';
+import { moveFlowInOrder, toggleInSet } from '../utils/flow-view-helpers.js';
 import {
   addCategory, renameCategoryInline, deleteCategory,
 } from '../utils/flow-view-categories.js';
@@ -99,8 +99,7 @@ export class FlowView extends ComponentBase {
   }
 
   _toggleCollapse(catId) {
-    if (this._collapsedCategories.has(catId)) this._collapsedCategories.delete(catId);
-    else this._collapsedCategories.add(catId);
+    toggleInSet(this._collapsedCategories, catId);
     this._renderList();
   }
