+++ src/components/settings-appearance.js
@@ -6,7 +6,7 @@ import { TERMINAL_THEMES, getTerminalThemeName, setTerminalTheme, getTerminalThe
 import { getAppTheme, setAppTheme } from '../utils/app-theme.js';
 import { _el, createActionButton } from '../utils/dom.js';
 import { MODE_BUTTONS, THEME_PREVIEW_LINES, COLOR_DOT_KEYS } from '../utils/settings-helpers.js';
-import { buildSettingsSection, createSettingsItem } from '../utils/settings-section-builder.js';
+import { buildSettingsSection, createListItem } from '../utils/settings-section-builder.js';
 import { registerComponent } from '../utils/component-registry.js';
 import { createAsyncHandler } from '../utils/event-helpers.js';
 
@@ -53,7 +53,7 @@ function _buildThemePreview(name, theme) {
 }
 
 function _createThemeCard(name, theme, isActive, tabManager, renderAppearanceFn) {
-  return createSettingsItem({
+  return createListItem({
     cls: 'theme-card',
     isActive,
     activeCls: 'theme-active',
