+++ src/utils/file-tree-section-dom.js
@@ -8,8 +8,7 @@
  * @typedef {{ setupDropZone: (el: HTMLElement, targetDir: string|(() => string|null)) => void, promptNewEntry: (dirPath: string, cEl: HTMLElement, depth: number, eDirs: Set<string>, type: string) => void, promptRename: (path: string, nameEl: HTMLElement) => void, refreshSection: (cwd: string) => void, contextMenuApi: unknown }} SectionDOMCallbacks
  */
 
-import { _el } from './dom.js';
-import { buildChevronRow } from './dom.js';
+import { _el, buildChevronRow, toggleCollapsible } from './dom.js';
 import { attachContextMenu } from './context-menu.js';
 import { emitWorkspaceCreateWorktree, emitWorkspaceOpenPr } from './workspace-events.js';
 import {
@@ -70,8 +69,7 @@ function _buildSectionHeader(cwd, contentEl, wasCollapsed, expandedDirs, callbac
   labelEl.title = cwd;
 
   header.addEventListener('click', () => {
-    const collapsed = contentEl.classList.toggle('collapsed');
-    chevron.textContent = collapsed ? CHEVRON_COLLAPSED : CHEVRON_EXPANDED;
+    toggleCollapsible(contentEl, chevron, CHEVRON_EXPANDED, CHEVRON_COLLAPSED);
   });
 
   attachContextMenu(header, () => buildDirContextItems(
